pyproject.toml
[project]
name = "ragdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.0",
    "gradio>=5.50.0",
    "jq>=1.10.0",
    "langchain>=1.0.8",
    "langchain-chroma",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "matplotlib>=3.10.7",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "sqlite3",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.20",
    "wikipedia>=1.4.0",
]
